+++ ThunderFocus-firmware/src/SunKeeper.h
@@ -10,14 +10,14 @@
 #include <SolarPosition.h>
 #define MIN_VALID_UNIX_TIME 1577836800
 
-extern float sunKeepLat;
-extern float sunKeepLong;
+extern double sunKeepLat;
+extern double sunKeepLong;
 
 void initTime();
-void setWorldCoord(float lat, float lng);
-float getWorldLat();
-float getWorldLong();
-float getSolarElevation();
+void setWorldCoord(double lat, double lng);
+double getWorldLat();
+double getWorldLong();
+double getSolarElevation();
 void setTeensyTime(unsigned long currentTime);
 time_t getTeensyTime();
